additional checks before the AppProxy and SysUpdate patch

+++ custom-script/custom_enable_local_ota.sh
@@ -37,35 +37,35 @@ function custom_parse_args_01_enable_local_ota() {
 
 function custom_function_01_enable_local_ota() {
     if [ $ENABLE_LOCAL_OTA -eq 1 ]; then
-        echo "+ Enable local ota"
-        echo "++ Trying to patch AppProxy"
-        cp "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy" "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy.xiaomi"
-        sed -i 's/\x7F\xF4\x29\xAF\x13\x79\x03\xF0\xDF\x03\x53\x2B\x7F\xF4\xCD\xAC\x21\xE7\x38\x46\xF0\xF7\xE2\xEB/\x7F\xF4\x29\xAF\xD3\x78\x03\xF0\xDF\x03\x50\x2B\x7F\xF4\xCD\xAC\x21\xE7\x38\x46\xF0\xF7\xE2\xEB/g' "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy"
-        MD5_ORG=$(md5sum "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy.xiaomi" | awk '{print $1}')
-        MD5_PATCHED=$(md5sum "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy" | awk '{print $1}')
-
-        if [ "$MD5_ORG" = "$MD5_PATCHED" ]; then
-            echo "-- AppProxy is NOT patched"
-            echo "!! Probably this firmware is not 2008+"
-        else
-            echo "++ AppProxy is patched"
-            echo "++ Trying to patch SysUpdate"
-            cp "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate" "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate.xiaomi"
-            sed -i 's/\x0B\x79\x03\xF0\xDF\x03\x53\x2B\x3E\xD0\x4C/\xCB\x78\x03\xF0\xDF\x03\x50\x2B\x3E\xD0\x4C/g' "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate"
-            MD5_ORG=$(md5sum "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate.xiaomi" | awk '{print $1}')
-            MD5_PATCHED=$(md5sum "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate" | awk '{print $1}')
+        if [ $FW_VER -ge 2008 -a $FW_VER -lt 3000 ]; then
+            echo "+ Enable local ota"
+            echo "++ Trying to patch AppProxy"
+            cp "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy" "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy.xiaomi"
+            sed -i 's/\x7F\xF4\x29\xAF\x13\x79\x03\xF0\xDF\x03\x53\x2B\x7F\xF4\xCD\xAC\x21\xE7\x38\x46\xF0\xF7\xE2\xEB/\x7F\xF4\x29\xAF\xD3\x78\x03\xF0\xDF\x03\x50\x2B\x7F\xF4\xCD\xAC\x21\xE7\x38\x46\xF0\xF7\xE2\xEB/g' "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy"
+            MD5_ORG=$(md5sum "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy.xiaomi" | awk '{print $1}')
+            MD5_PATCHED=$(md5sum "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy" | awk '{print $1}')
 
             if [ "$MD5_ORG" = "$MD5_PATCHED" ]; then
-                echo "-- SysUpdate is NOT patched"
-                echo "!! Restore original AppProxy"
-                echo "!! Probably this firmware is not 2008+"
-                cp "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy.xiaomi" "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy"
+                echo "-- AppProxy is NOT patched"
             else
-            echo "++ SysUpdate is patched"
+                echo "++ AppProxy is patched"
+                echo "++ Trying to patch SysUpdate"
+                cp "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate" "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate.xiaomi"
+                sed -i 's/\x0B\x79\x03\xF0\xDF\x03\x53\x2B\x3E\xD0\x4C/\xCB\x78\x03\xF0\xDF\x03\x50\x2B\x3E\xD0\x4C/g' "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate"
+                MD5_ORG=$(md5sum "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate.xiaomi" | awk '{print $1}')
+                MD5_PATCHED=$(md5sum "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate" | awk '{print $1}')
+
+                if [ "$MD5_ORG" = "$MD5_PATCHED" ]; then
+                    echo "-- SysUpdate is NOT patched"
+                    echo "!! Restore original AppProxy"
+                    cp "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy.xiaomi" "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy"
+                else
+                echo "++ SysUpdate is patched"
+                fi
             fi
+
+            rm -f "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy.xiaomi"
+            rm -f "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate.xiaomi"
         fi
     fi
-
-    rm -f "${IMG_DIR}/opt/rockrobo/cleaner/bin/AppProxy.xiaomi"
-    rm -f "${IMG_DIR}/opt/rockrobo/cleaner/bin/SysUpdate.xiaomi"
 }
